Connected GetEventStoreRepository to The Aggregate Source IEventSouce Fixed all Tests Implemented Embbeded Eventstore Need to Add Perfix Suffix Support

+++ src/ReactiveDomain.Foundation/EventStore/AggregateVersionException.cs
@@ -8,10 +8,10 @@ namespace ReactiveDomain.Foundation.EventStore
     {
         public readonly Guid Id;
         public readonly Type Type;
-        public readonly int AggregateVersion;
-        public readonly int RequestedVersion;
+        public readonly long AggregateVersion;
+        public readonly long RequestedVersion;
         
-        public AggregateVersionException(Guid id, Type type, int aggregateVersion, int requestedVersion)
+        public AggregateVersionException(Guid id, Type type, long aggregateVersion, long requestedVersion)
             : base(string.Format("Requested version {2} of aggregate '{0}' (type {1}) - aggregate version is {3}", id, type.Name, requestedVersion, aggregateVersion))
         {
             Id = id;
